+++ source/B2B.Transactions/Xml/MessageValidator.cs
@@ -21,7 +21,7 @@ using System.Xml;
 using System.Xml.Schema;
 using B2B.Transactions.Xml.Incoming;
 
-namespace B2B.Transactions.Xml.Outgoing
+namespace B2B.Transactions.Xml
 {
     public class MessageValidator
     {
